data-control: Set O_NONBLOCK on the receive and send fd
A misbehaving client can cause us to block when writing to the send fd,
such as if they create a pipe, read only a few bytes, then block
themselves, but we called write() with megabytes of data.

As for the receive fd, the Linux man page for select(2) says there may
be "circumstances in which a file descriptor is spuriously reported as
ready". This is probably never going to be a problem and we'd use epoll
on Linux anyway, but setting it doesn't hurt.

(O_NONBLOCK helps if the misbehaving client uses a pipe, but writing to
a disk would still "block" (more accurately, put us in non-interruptible
sleep). The protocol doesn't specify a limitation on what kinds of files
can be used, so there are probably various ways a client can freeze the
selection owner.)