data-control: Fallback to asynchronous source send
This prevents sending incomplete data if the system or receiving Wayland
client causes partial writes to occur (especially now that we set
O_NONBLOCK). This needs to be asynchronous (not a while loop) to protect
us from misbehaving clients.

We only need to perform asynchronous sending if a partial write occurs
on the initial synchronous write. Otherwise, don't bother. That means we
should keep O_NONBLOCK for the initial synchronous write.