main: Allow websockets to use fd and unix sockets
The docs for the `--unix-socket` and `--external-listener-fd` options
sound like they should work with the `--websocket` option. This change
implements that.

By allowing wayvnc to accept an arbitrary FD to listen on, we can use
systemd's socket activation functionality to start wayvnc on demand.

I am adding this as I want to put wayvnc behind nginx, and only start
wayvnc when a webclient is trying to connect to the backend.