client-resize: Stop rate_limiter timer
A VNC client resize request can be received when the rate limiter timer
is active. If the timer fires between on_output_dimension_change()
calling wayvnc_start_capture_immediate() and the capture finishing,
wayvnc_start_capture() is called with a capture already in progress.

This is fatal when the wlr_screencopy_unstable_v1 interface is being used.